Project import generated by Copybara.
PiperOrigin-RevId: 200741825

+++ datathon/organizer/scripts/upload_data.sh
@@ -68,19 +68,28 @@ if [[ -z ${OWNERS_GROUP} ]] || [[ -z ${EDITORS_GROUP} ]] || \
   exit 1
 fi
 
-# Lowercase is required for bucket name and deployment job name.
-DATASET_NAME=`echo ${DATASET_NAME} | awk '{ print tolower($0) }'`
+# Lowercase is required for bucket name, and BigQuery dataset ID can't contain
+# dash.
+DATASET_NAME=`echo ${DATASET_NAME} | awk '{ print tolower($0) }' \
+  | sed -e 's/-/_/g'`
 BUCKET_ID=${PROJECT_ID}-${DATASET_NAME}
 DATASET_ID=${DATASET_NAME}
-LOCATION=US
+LOCATION=US More flexible state config.
+STATE_FILE="$0".state
+# A list of state checkpoints for the script to resume to.
+STATE_ENABLE_SERVICES="ENABLE_SERVICES"
+STATE_CREATE_BQ="CREATE_BQ"
+STATE_UPLOAD_GCS="UPLOAD_GCS"
+STATE_UPLOAD_BQ="UPLOAD_BQ"
 
-echo "Creating the Google Cloud Storage bucket to host data."
-STORAGE_CLASS=multi_regional
-gsutil mb -p ${PROJECT_ID} -c ${STORAGE_CLASS} -l ${LOCATION} gs://${BUCKET_ID}
-gsutil requesterpays set on gs://${BUCKET_ID}
-echo "Setting Google Cloud Storage bucket access."
-TEMP=`tempfile`
-cat <<EOF >>${TEMP}
+if [[ ! -e ${STATE_FILE} ]]; then
+  echo "Creating the Google Cloud Storage bucket to host data."
+  STORAGE_CLASS=multi_regional
+  gsutil mb -p ${PROJECT_ID} -c ${STORAGE_CLASS} -l ${LOCATION} gs://${BUCKET_ID}
+  gsutil requesterpays set on gs://${BUCKET_ID}
+  echo "Setting Google Cloud Storage bucket access."
+  TEMP=`tempfile`
+  cat <<EOF >>${TEMP}
 {
   "bindings": [
     {
@@ -104,16 +113,26 @@ cat <<EOF >>${TEMP}
   ]
 }
 EOF
-gsutil iam set ${TEMP} gs://${BUCKET_ID}
+  gsutil iam set ${TEMP} gs://${BUCKET_ID}
+  echo ${STATE_ENABLE_SERVICES} > ${STATE_FILE}
+else
+  echo "Skip creating bucket since it has previously finished."
+fi
 
-echo "Enabling the following Google Cloud Platform services"
-echo "  - Deployment Manager"
-echo "  - BigQuery"
-gcloud services enable deploymentmanager bigquery --project ${PROJECT_ID}
+if [[ `cat ${STATE_FILE}` == ${STATE_ENABLE_SERVICES} ]]; then
+  echo "Enabling the following Google Cloud Platform services"
+  echo "  - Deployment Manager"
+  echo "  - BigQuery"
+  gcloud services enable deploymentmanager bigquery --project ${PROJECT_ID}
+  echo ${STATE_CREATE_BQ} > ${STATE_FILE}
+else
+  echo "Skip enabling services since it has previously finished."
+fi
 
-echo "Creating the BigQuery to host tables."
-TEMP=`tempfile`
-cat <<EOF >>${TEMP}
+if [[ `cat ${STATE_FILE}` == ${STATE_CREATE_BQ} ]]; then
+  echo "Creating the BigQuery to host tables."
+  TEMP=`tempfile`
+  cat <<EOF >>${TEMP}
 resources:
 - name: big-query-dataset
   type: bigquery.v2.dataset
@@ -129,15 +148,24 @@ resources:
         groupByEmail: "${READERS_GROUP?}"
     location: "${LOCATION?}"
 EOF
-gcloud deployment-manager deployments create create-${DATASET_ID} \
-  --config=${TEMP} --project ${PROJECT_ID}
-# Delete the deployment job since it's finished. The created dataset is not
-# affected.
-gcloud --quiet deployment-manager deployments delete create-${DATASET_ID} \
-  --project ${PROJECT_ID} --delete-policy=ABANDON
+  gcloud deployment-manager deployments create create-bq-dataset \
+    --config=${TEMP} --project ${PROJECT_ID}
+  # Delete the deployment job since it's finished. The created dataset is not
+  # affected.
+  gcloud --quiet deployment-manager deployments delete create-bq-dataset \
+    --project ${PROJECT_ID} --delete-policy=ABANDON
+  echo ${STATE_UPLOAD_GCS} > ${STATE_FILE}
+else
+  echo "Skip creating BigQuery dataset since it has previously finished."
+fi
 
-echo "Uploading data to Google Cloud Storage."
-gsutil -m cp ${INPUT_DIR}/*.csv.gz gs://${BUCKET_ID}
+if [[ `cat ${STATE_FILE}` == ${STATE_UPLOAD_GCS} ]]; then
+  echo "Uploading data to Google Cloud Storage."
+  gsutil -m cp ${INPUT_DIR}/*.csv.gz gs://${BUCKET_ID}
+  echo ${STATE_UPLOAD_BQ} > ${STATE_FILE}
+else
+  echo "Skip uploading data to GCS since it has previously finished."
+fi
 
 GOEXEC=`which go`
 BQ_SCHEMA_DETECTOR_PATH=`dirname "$0"`/../src/bq/bq_schema_detector.go
@@ -157,42 +185,48 @@ generate_schema() {
   echo ${schema_file}
 }
 
-job_ids=()
-echo "Loading data from Google Cloud Storage to BigQuery."
-for file in ${INPUT_DIR}/*.csv.gz; do
-  echo "Loading data file `basename ${file}` to BigQuery."
-  # Assume the schema file ends with .csv.gz.schema suffix.
-  schema_file=${SCHEMA_DIR}/$(basename ${file}).schema
-  delete_schema_file=false
-  if [[ ! -e ${schema_file} ]]; then
-    # Generate the schema file if it does not already exist.
-    echo "Generating schema for ${file}."
-    schema_file=`generate_schema ${file}`
-    delete_schema_file=true
-  fi
+if [[ `cat ${STATE_FILE}` == ${STATE_UPLOAD_BQ} ]]; then
+  job_ids=()
+  echo "Loading data from Google Cloud Storage to BigQuery."
+  for file in ${INPUT_DIR}/*.csv.gz; do
+    echo "Loading data file `basename ${file}` to BigQuery."
+    # Assume the schema file ends with .csv.gz.schema suffix.
+    schema_file=${SCHEMA_DIR}/$(basename ${file}).schema
+    delete_schema_file=false
+    if [[ ! -e ${schema_file} ]]; then
+      # Generate the schema file if it does not already exist.
+      echo "Generating schema for ${file}."
+      schema_file=`generate_schema ${file}`
+      delete_schema_file=true
+    fi
 
-  # Upload data to BigQuery.
-  table_name=`echo $(basename ${file:0:(-7)}) | awk '{ print tolower($0) }'`
+    # Upload data to BigQuery.
+    table_name=`echo $(basename ${file:0:(-7)}) | awk '{ print tolower($0) }'`
 
-  # Asynchronously load data to BigQuery.
-  job_id=${table_name}_$(date +%s)
-  bq --nosync --location=${LOCATION} --project_id=${PROJECT_ID} load \
-    --job_id=${job_id} --skip_leading_rows=1 --source_format=CSV \
-    --allow_quoted_newline "${DATASET_ID}.${table_name}" \
-    gs://${BUCKET_ID}/$(basename ${file}) ${schema_file}
-  job_ids+=(${job_id})
+    # Asynchronously load data to BigQuery.
+    job_id=${table_name}_$(date +%s)
+    bq --nosync --location=${LOCATION} --project_id=${PROJECT_ID} load \
+      --job_id=${job_id} --skip_leading_rows=1 --source_format=CSV \
+      --allow_quoted_newline "${DATASET_ID}.${table_name}" \
+      gs://${BUCKET_ID}/$(basename ${file}) ${schema_file}
+    job_ids+=(${job_id})
 
-  # Remove the schema file if it's generated by this script.
-  if [[ ${delete_schema_file} = true ]]; then
-    rm ${schema_file}
-  fi
-done
+    # Remove the schema file if it's generated by this script.
+    if [[ ${delete_schema_file} = true ]]; then
+      rm ${schema_file}
+    fi
+  done
 
-echo "All BigQuery loading jobs have been submitted."
-echo "Waiting for jobs to finish."
-for job_id in ${job_ids[*]}; do
-  # The wait command will print out real-time status of the job.
-  bq wait --project_id=${PROJECT_ID} ${job_id}
-done
+  echo "All BigQuery loading jobs have been submitted."
+  echo "Waiting for jobs to finish."
+  for job_id in ${job_ids[*]}; do
+    # The wait command will print out real-time status of the job.
+    bq wait --project_id=${PROJECT_ID} ${job_id}
+  done
+else
+  echo "Skip uploading data to BigQuery since it has previously finished."
+fi
+
+rm ${STATE_FILE}
 
 echo "Data upload for ${DATASET_NAME} has finished."
